git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_2_BRANCH@9184 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/hash.h
@@ -244,7 +244,7 @@ private:
     protected:                                                                 \
         void DoPut(long key, long value, eltype *data)                         \
         {                                                                      \
-            size_t slot = (size_t)abs(key % m_hashSize);                       \
+            size_t slot = (size_t)abs((int)(key % (long)m_hashSize));          \
                                                                                \
             if ( !m_hashTable[slot] )                                          \
             {                                                                  \
